// src/routes/router.type.ts
export enum RouterPath {
  REACT_COMPILER = "/react-compiler-example",
  USE_EXAMPLE = "/use-example",
  USE_EXAMPLE_2 = "/use-example-2",
  ACTION_EXAMPLE = "/action-example",
  USE_FORM_STATE = "/useFormState",
  USE_FORM_STATUS = "/useFormStatus",
  USE_OPTIMISTIC = "/useOptimistic",
  DOCUMENT_METADATA = "/documentMetadata",
  REF_HANDLING = "/refHandling",
}